Examples: Move webgl_worker_offscreencanvas to modules.
